Fix preview background threading, add thread pool of size on, with unbounded work queue
Argument to (Thread.) call had one parenthesis too much, so preview
creation was invoked in the calling thread...

Long running preview creation still has a problem since it relies on
the upload file to be present. Uploaded files are removed after the
request completes, see:

http://beelse.solita.fi/hg/lupapalvelu/file/94636c13076e/src/lupapalvelu/web.clj#l458

The uploaded file should be copied, and remove after the submitted work
is done.