create new autocomplete component, which is simplifies the use of the base-autocomplete component
this autocomplete component embeds the filtering of options in autocomplete UI (user's query string search), which previously needed to be implemented for each autocomplete use case.
The simplified conventions are documented to docstring in autocomplete-model.js. Basic idea is to give array of objects which act as selectable options. Autocomplete UI is used to select one of those object's as the value (objects' display property can be set).