Verify that attachment id given as parameter in file upload is valid
If the user provides an attachment id, it must mean that the given
attachment is to be updated. Previously the user could provide an id
and this would be used to create a new attachment if an existing with
that id was not found under the current application. Now the upload
fails in this case.