set notNeeded to false when creating version for attachment
If file is uploaded, it is needed (not "not needed") then. Uploading to notNeeded attachments is blocked with pre-checks, but copy-user-attachments command resulted in versions uploaded to possibly not needed attachments.
Alternative solution would have been to filter out notNeeded attachment placeholders, but that approach wasn't bullet proof as user-attachment use special id:s. This resulted new version in attachment, even though it was marked as notNeeded and was filtered in "same attachment type" matching.