Fix ya-extension-itest and related robot tests
This is a bit of a dirty hack, as the problem is that the sample XML
had application number tied to 2016 while the test applications were
created dynamically, i.e. receiving the current year in their id. Now
we replace the year placeholder in the XML with the current year before
parsing it.