Prevent exception in String/format by using the non-format version of timbre debug
It seems that the debugf call in sade.util/timing sometimes caused
an exception in String/format:

IllegalFormatConversionException: c != java.lang.Long

It's not clear why the argument sometimes is not Long and sometimes is,
but after this commit we only require it to be something that can be
output as String. Hopefully this fixes the exceptions witnessed in
production.