select distinct reviews from XML after dissociating muuTunnus
If review is sent from Lupapiste to backingsystem (ie KuntaNet), it might receive new ID in receiving end.
When Lupapiste reads data back, the data could be duplicated in KuntaGML XML message, but with different muuTunnus values.
By dissociating muuTunnus and comparing the data itself, we can be as sure as possible, that duplicates are removed by comparing the *actual* data. (taskname, pitoPvm, huomautukset, lasnaolijat etc)