Improve Robot Chrome compatibility with custom Selenium initialization and scrolling
Chrome must have clickable items visible in actual viewport to be able
to click them. Thus the element should be scrolled into view before
trying to click it. Chrome will also throw an error if it tries to
click on an element that is not clickable (i.e. active button / link).