fix change_permit robot to mimic more realistic scenario
Old test had application created by authority, and thus after new permissions model, the co-created change-permit was initally in "wrong" state (draft instead of open).
But it is more realistic to create change-permit based on actual applicant's application. This way the state is actually 'open' because authority is not initially.

Also added new fixture for Robot + smoketests for all fixtures