LPK-3683 replace property-id->municipality resolving by querying KTJKii for municipality information
Previous logic in sade.property was to deduce municipality from propertyId string. Possible municipality merges was handled by mapping in sade.municipality ns. But this parsing from string is potentially unreliable.
For example if arbitrary area of municipality Y is transferred to municipality X, then it's impossible to deduce correct municipality with the propertyId and mapping table (case Jakkukylä was transferred from Oulu to Ii).