LPK-3694 move also designers that are approved, after verdict has been given
This enables authorities to move designer even if for some reason it was not approved in the beginning of the project, but later after verdict. Before this change the only way to transfer those was to create totally new document and fill details again (as predicate was 'created' timestamp instead of approval).