refactor ramitest to use bind-attachments
notice that pohjapiirustus was changed to julkisivupiirustus, because behaviour changed when we changed upload command.
Previously new attachment was created, but bind-attachments checks if there are empty attachments and places the file to empty placeholder. Thus the attachment ID's did not match in test, so it was easier to put totally new type to ensure test runs as previously.