LP-366498 if 'address' is not present in KuntaGML, but other information is sufficient, resolve address from NLS
Previously requests were returned with (fail :error.more-prev-app-info-needed) if address was missing. This commit fixes this situation if we have the more important information already at hand: x,y and propertyId. Address is just "nice to have", so it was bad that it returned error.code when we had all the possibilities to query for some address (x+y -> address)