LPK-3689 statementGiver can call update-user
and (organization) statementGivers are "authority" roles

The reason this command was not previously on allowed-actions was that the command contained `:user-roles` key, which denoted that the command will use default-writer-authz-roles as `:user-authz-roles` for commands, and thus statementGiver role was out of the question.
Now new `update-user` command has only permissions, so the actual `:user-org-authz` are much wider for that command.

But it makes sense that statementGiver can update own info :)

Nevertheless, test is a bit funny because it gives application as parameter, even though many of the tested actions do not require application as parameters, as they are more generic commands (like update-user)...