Remove deprecated $isolated from Mongo queries.
$isolated provided full database lock to the thread doing Mongo updates,
thus ensuring that any readers would not see changes in only some of
the documents being updated. But as the updates would always be only
to a single collection and we do not have documents referencing other
documents in same collection nor do we care if someone reads only some
updated docs, this option should be safe to remove.