Add placeholder like functionality to autocomplete-component
When nothing is selected, we might want to show something (like call to
action) in component. Placeholder option does this.

Component seems to handle empty string as an value that means nothing
has been selected.

Alternative option would be to handle this in items and callbacks, but I
think that would basially add a "null-check" variant to calling code...