Add preliminary version of invoice update and insert
Currently an upsert function in frontend. We still have little
difficulties on how to actually work with new invoice.

It get's saved when operation is added, and is returned to invoice list
like others. But we need a way to sort the stuff, so that it would be
the first one.