LPK-3433 Invoice saving and updating works from UI
* Invoice component is updated in place.
Invoice atom is used in all components that update state. The sillyside
is that the initial vector of fetched invoices is pretty much obsolete
after first edit (that doesn't trigger new fetch).

* When invoice-row has type :custom, it can be edited quite freely.

* State management for invoice (draft, complete) is pretty simple for
now, and a little buggy. We need a data structure for
states and control flow.

* Populate operation add dropdown from data fetched from backend.
Contains operations that are in the application.

* Sort invoices by draft state and then :created timestamp