Use query parameter data from (:params ring-req) instead of (:query-params ring-req) so that nested parameter handling is supported
When there is a sequential query param (=list) in the request in the form
http://foobar?listOfStuff[]=A&listOfStuff[]=B
it is converted to a vector in the :params map like this
{:params {:listOfStuff ["A" "B"] ...}

And a single item list
http://foobar?listOfStuff[]=753-R -> {:params {:listOfStuff ["A"] ...}

This is done probably by a middleware like ring.wrap-nested-params somewhere in the noir code.
However, this handling does not affect the :query-params map.
Therefore we need to use the :params instead of the :query-params from the ring request.