LPK-3766 Assignments and handlers.
Handler upserts are reflected in the assignments. An automatic
assignment always has the current corresponding handler as its
recipient. The recipient is empty (or rather includes only the roleId)
if no suitable handler is defined.

Modified timestamp for assignments. Also a corresponding column in the
UI.

Refresh application assignments when handlers have been edited.