Removes debug prints and fixes a tricky bug where a late lazy seq computation caused deletion of wrong catalogue
When a call to delete-catalogues for catalogues with the same valid-from
was after creation of the new catalogue. The new catalogue was deleted
as well because it had the same valid-from day even though the
collection of deletable catalogues was created earlier. This is because
the collection had not been computed yet and when the colletion was
first used, the newly created catalogue with its same-valid from was now
part of the collection of deletable catalogues.