TT-346 Reset attachments before fetching them from backend
When tos-function is updated, the original idea is to load and reset
attachments as well.

For some reason, knockout forgets the necessary bindings for backend
communication.

But, if we empty the attachments list before populating it again, the
screen re-render triggers and communication with backend starts to work.

In theory, this should not cause problems, since I take it that the
reload function IS ment to update all attachments.